Refactor code and attempt to update the CommonLib so that it works for everyone.
