+++ cmake/version.rc.in
@@ -1,0 +1,32 @@
+#include <winres.h>
+
+1 VERSIONINFO
+	FILEVERSION @PROJECT_VERSION_MAJOR@, @PROJECT_VERSION_MINOR@, @PROJECT_VERSION_PATCH@, 0
+	PRODUCTVERSION @PROJECT_VERSION_MAJOR@, @PROJECT_VERSION_MINOR@, @PROJECT_VERSION_PATCH@, 0
+	FILEFLAGSMASK 0x17L
+#ifdef _DEBUG
+	FILEFLAGS 0x1L
+#else
+	FILEFLAGS 0x0L
+#endif
+	FILEOS 0x4L
+	FILETYPE 0x1L
+	FILESUBTYPE 0x0L
+BEGIN
+	BLOCK "StringFileInfo"
+	BEGIN
+		BLOCK "040904b0"
+		BEGIN
+			VALUE "FileDescription", "@PROJECT_DESCRIPTION@"
+			VALUE "FileVersion", "@PROJECT_VERSION@"
+			VALUE "InternalName", "@PROJECT_NAME@"
+			VALUE "LegalCopyright", "Apache License 2.0"
+			VALUE "ProductName", "@PROJECT_FRIENDLY_NAME@"
+			VALUE "ProductVersion", "@PROJECT_VERSION@"
+		END
+	END
+	BLOCK "VarFileInfo"
+	BEGIN
+		VALUE "Translation", 0x409, 1200
+	END
+END
