+++ include/DisguiseData.h
