+++ include/NpcDetectionData.h
