--- include/log.h
@@ -1,6 +1,0 @@
-#include <filesystem>
-#include <fstream>
-#include <iostream>
-#include <string>
-
-void LogToFile(const std::string& message);
