+++ src/ArmorKeywordData.cpp
@@ -1,5 +1,5 @@
-#include "armor_keyword_data.h"
-#include "armor_slots.h"
+#include "ArmorKeywordData.h"
+#include "ArmorSlots.h"
 
 
 void SaveArmorKeywordDataCallback(SKSE::SerializationInterface* a_intfc) {
